backend/requirements.txt
flask
flask_cors
sqlalchemy
psycopg2-binary
langchain
openai
langchain-community
python-dotenv
langchain-openai